drakorhoverpad_updateMain: s16 yawDelta narrows Obj_GetYawDeltaToObject result at assign (extsh after call), clamp works in s16 reg (99.674->100, unit->100)

+++ src/main/dll/dll_0271_drakorhoverpad.c
@@ -519,7 +519,7 @@ void drakorhoverpad_updateMain(GameObject* obj)
     f32 absH;
     f32 absV;
     GameObject* nearest;
-    int yawDelta;
+    s16 yawDelta;
     int c;
     int angle;
     int clamped;
